Refactor: Extract export generation logic to ExportService
